refactor(article-search): split providers ranking and parsing
